# IGNITE-22703 Sql. Low performance for select with mixed projection