--- a/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/TableRowConverterFactory.java
+++ b/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/TableRowConverterFactory.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.sql.engine.exec;
 
-import java.util.BitSet;
+import org.apache.calcite.util.ImmutableIntList;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -26,9 +26,9 @@ import org.jetbrains.annotations.Nullable;
  */
 @FunctionalInterface
 public interface TableRowConverterFactory {
-    TableRowConverter create(@Nullable BitSet requiredColumns);
+    TableRowConverter create(@Nullable ImmutableIntList projection);
 
-    default TableRowConverter create(BitSet requiredColumns, int partId) {
-        return create(requiredColumns);
+    default TableRowConverter create(@Nullable ImmutableIntList projection, int partId) {
+        return create(projection);
     }
 }