--- a/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/TableRowConverterFactoryImpl.java
+++ b/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/TableRowConverterFactoryImpl.java
@@ -96,7 +96,6 @@ public class TableRowConverterFactoryImpl implements TableRowConverterFactory {
                 ? tableColumnSet
                 : projection;
 
-        // TODO: IGNITE-22703 Can we opmitimize this?
         if (Commons.isIdentityMapping(mapping, schemaDescriptor.length())) {
             return fullRowConverter;
         }