--- a/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/rel/TableScanNode.java
+++ b/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/rel/TableScanNode.java
@@ -46,7 +46,7 @@ public class TableScanNode<RowT> extends StorageScanNode<RowT> {
 
     private final RowFactory<RowT> rowFactory;
 
-    private final @Nullable int[] requiredColumns;
+    private final int @Nullable [] requiredColumns;
 
     /**
      * Constructor.