@AMashenkov 2025-07-04
Before the fix
```
Benchmark                  (fsync)  (projection)  Mode  Cnt    Score   Error  Units
SqlSelectAllBenchmark.run    false      IDENTITY  avgt   20  170.579 ± 0.185  ms/op
SqlSelectAllBenchmark.run    false      TRIMMING  avgt   20  335.625 ± 0.329  ms/op
SqlSelectAllBenchmark.run    false   RESHUFFLING  avgt   20  564.525 ± 0.692  ms/op
```
After the fix
```
Benchmark                  (fsync)  (projection)  Mode  Cnt    Score   Error  Units
SqlSelectAllBenchmark.run    false      IDENTITY  avgt   20  173.467 ± 0.153  ms/op
SqlSelectAllBenchmark.run    false      TRIMMING  avgt   20  344.805 ± 0.378  ms/op
SqlSelectAllBenchmark.run    false   RESHUFFLING  avgt   20  348.435 ± 0.283  ms/op

```